+++ src/providers/SidebarProvider.ts
@@ -8,12 +8,12 @@ import {
 } from "vscode";
 import { getUri } from "../utilities/getUri";
 import { getNonce } from "../utilities/getNonce";
-import { SidebarEventListener } from "../listener/SidebarEventListener";
+import { EventListener } from "../listener/SidebarEventListener";
 
-export class SidebarProvider implements WebviewViewProvider {
-  public static readonly viewType = "voice-tweet-explorer";
+export class ViewProvider implements WebviewViewProvider {
+  public static readonly viewType = "voice-tweet";
 
-  constructor(private readonly _extensionUri: Uri) {}
+  constructor(private readonly _extensionUri: Uri) { }
 
   public resolveWebviewView(
     webviewView: WebviewView,
@@ -27,12 +27,12 @@ export class SidebarProvider implements WebviewViewProvider {
 
     webviewView.webview.html = this._getWebviewContent(webviewView.webview, this._extensionUri);
 
-    const listener = new SidebarEventListener();
+    const listener = new EventListener();
     listener.setWebviewMessageListener(webviewView);
   }
 
   private _getWebviewContent(webview: Webview, extensionUri: Uri) {
-    const webviewUri = getUri(webview, extensionUri, ["out", "sidebar.js"]);
+    const webviewUri = getUri(webview, extensionUri, ["out", "webview.js"]);
     const stylesUri = getUri(webview, extensionUri, ["out", "styles.css"]);
     const nonce = getNonce();
